Ver3/requirements.txt
asgiref==3.5.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
numpy==1.23.5
paho-mqtt==1.6.1
pandas==1.5.2
psycopg2==2.9.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.30.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.3
timedelta==2020.12.3
Twisted==22.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2022.4
urllib3==2.0.2
zope.interface==6.0
